% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pad_geoid}
\alias{pad_geoid}
\title{Pad GEOID to 5 digits}
\usage{
pad_geoid(geoid)
}
\arguments{
\item{geoid}{Character or numeric vector of geographic identifiers}
}
\value{
Character vector of 5-digit zero-padded GEOIDs
}
\description{
Ensures geographic identifiers are zero-padded to 5 digits.
}
\examples{
pad_geoid(c("123", "45678", 789))
#> [1] "00123" "45678" "00789"
}
