% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{print.shellgame_audit}
\alias{print.shellgame_audit}
\title{Print method for shellgame_audit}
\usage{
\method{print}{shellgame_audit}(x, ...)
}
\arguments{
\item{x}{A shellgame_audit object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for shellgame_audit
}
