% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{summary.shellgame_audit}
\alias{summary.shellgame_audit}
\title{Summary method for shellgame_audit}
\usage{
\method{summary}{shellgame_audit}(object, ...)
}
\arguments{
\item{object}{A shellgame_audit object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary method for shellgame_audit
}
