% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_columns}
\alias{validate_columns}
\title{Validate required columns in data frame}
\usage{
validate_columns(data, required_cols, data_name = "data")
}
\arguments{
\item{data}{Data frame to validate}

\item{required_cols}{Character vector of required column names}

\item{data_name}{Name of the data object (for error messages)}
}
\value{
Invisible TRUE if valid, stops with error if not
}
\description{
Validate required columns in data frame
}
\keyword{internal}
