% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{extract_perturbed_population}
\alias{extract_perturbed_population}
\title{Extract perturbed population details}
\usage{
extract_perturbed_population(audit_result, top_n = 10)
}
\arguments{
\item{audit_result}{A shellgame_audit object}

\item{top_n}{Number of top counties to return (default: 10)}
}
\value{
Data frame of counties that received population from target county
}
\description{
Extract perturbed population details
}
